[GFC] Move PlacedGridItem construction from ImplicitGrid to GridFormattingContext. https://bugs.webkit.org/show_bug.cgi?id=299533 rdar://problem/161337846
Reviewed by Brent Fulgham.

Currently, PlacedGridItems are created via placedGridItems() on
ImplicitGrid by constructing them out of the set of UnplacedGridItems
which it has mapped to different areas in the grid.

This has worked well enough for now, but as we prepare to size the
grid items, we need to be able to store additional data on these objects
in order to accomplish this. In particular, we need to store and be able
to use the various sizing properties (width, min-width, max-width, etc.)
for each dimension (inline vs block). However, we do not have enough
information at the ImplicitGrid level in order to map these properties
to the appropriate dimension. Instead, we should have GridFormattingContext
start creating these objects.

Not only will this allow us to correctly create PlacedGridItems with the
data that we need, but it also allows ImplicitGrid to return some simple
structures that describe the ImplicitGrid rather than a more complex
object.

* Source/WebCore/layout/formattingContexts/grid/GridLayout.cpp:
(WebCore::Layout::GridLayout::placeGridItems):
(WebCore::Layout::GridLayout::layout):
Previously we were getting the PlacedGridItems as a result of the grid
item placement portion (placeGridItems()) of the layout algorithm. Now
we get the grid areas for each item which were resolved from that
process. We can then ask the GridFormattingContext to create the set of
PlacedGridItems from this data.

* Source/WebCore/layout/formattingContexts/grid/ImplicitGrid.cpp:
(WebCore::Layout::ImplicitGrid::gridAreas const):
Instead of returning PlacedGridItems we return the set of grid areas
contained within the ImplicitGrid. This is just a map of each item that
has been placed in the grid along with the area (set of lines) that it
has been placed in. The caller can then take the UnplacedGridItem and
its grid area and do whatever it needs to which includes constructing a
PlacedGridItem if they need it.

Canonical link: https://commits.webkit.org/300547@main
