+++ Source/WebCore/layout/formattingContexts/grid/GridFormattingContext.h
@@ -33,8 +33,11 @@ namespace WebCore {
 namespace Layout {
 
 class ElementBox;
+class PlacedGridItem;
 
 class UnplacedGridItem;
+
+struct GridAreaLines;
 struct UnplacedGridItems;
 
 class GridFormattingContext : public CanMakeCheckedPtr<GridFormattingContext> {
@@ -50,6 +53,10 @@ public:
     GridFormattingContext(const ElementBox& gridBox, LayoutState&);
 
     void layout(GridLayoutConstraints);
+
+    using PlacedGridItems = Vector<PlacedGridItem>;
+    using GridAreas = HashMap<UnplacedGridItem, GridAreaLines>;
+    PlacedGridItems constructPlacedGridItems(const GridAreas&) const;
 
     const ElementBox& root() const { return m_gridBox; }
